Re-arranged code into a python module of sorts.  The \'flint\' folder now contains most of the provider-specific and module-specific code as well as some supporting utility functions.  Right now, ec2 is the only provider, and Spark is the only module. This is not meant to be a final code layout so much as a strawman proposal, and a first attempt at teasing apart the EC2 code from the rest of flintlock so that the GCE interface can be implemented cleanly in a parallel fashion.  Some things that are likely to change include (1) the name of the module, which is currently just "flint"; (2) the class or function structure of a provider... likely a class would be a better container for these functions at it would make it easier to track cluster state.  Nick\'s comments above the ClusterInfo class indicate he has had similar thoughts; (3) the exact naming and layout of the supporting files.  For now I have set aside ssh.py and util.py to contain functions that are commonly used throughout the code.
