+++ flintrock
@@ -1,3 +1,3 @@
 #!/usr/bin/env sh
 
-venv/bin/python3 -Wdefault::DeprecationWarning flintrock.py "$@"
+env python3 -Wdefault::DeprecationWarning flintrock.py "$@"
